+++ CHANGELOG.md
@@ -6,6 +6,28 @@ All user visible changes to this project will be documented in this file. This p
 
 
 
+## [0.12.0] · unreleased
+[0.12.0]: https://github.com/instrumentisto/medea-turn-rs/tree/v0.12.0
+
+[Diff](https://github.com/instrumentisto/medea-turn-rs/compare/v0.11.2...v0.12.0)
+
+### BC Breaks
+
+- Split `Config` to `Config` and `TurnConfig` for [TURN] optionality. ([#8])
+- Changed `Transport::send_to` to accept `Cow<[u8]>` instead of `Vec<u8>`. ([#8])
+- Added `transport::Error::Encode` variant. ([#8])
+- Added `chandata::FormatError::BufferTooShort` variant. ([#8])
+
+### Added
+
+- `Transport::send_msg_to()` method. ([#8])
+- `NoneAuthHandler` for use in type signatures when [TURN] is disabled. ([#8])
+
+[#8]: https://github.com/instrumentisto/medea-turn-rs/pull/8
+
+
+
+
 ## [0.11.2] · 2025-06-21
 [0.11.2]: https://github.com/instrumentisto/medea-turn-rs/tree/v0.11.2